Completed 'compact' map experiment. We were able to cut the size of the occupancy map by nearly 99.9% but SLAM latency doubled.

+++ MapIndex.h
@@ -5,8 +5,8 @@
 
 struct MapIndex
 {
-    uint32_t offset;
-    uint32_t size;
+    uint8_t offset;
+    uint8_t size;
 };
 
 #endif
